+++ install-lib.sh
@@ -23,6 +23,8 @@ mkdir -p $DESTDIR/usr/lib
 
 cp ./build/libinstaller.so.0.4.0 $DESTDIR/usr/lib/
 cd $DESTDIR/usr/lib/
+if [ ! -f libinstaller.so.0.4 ]; then
 ln -s /usr/lib/libinstaller.so.0.4.0 libinstaller.so.0.4
+fi
 
 echo "Installation of libinstaller finished."
